Move expand and block functionalities to Jacobian
