+++ src/Jacobian.cpp
@@ -434,6 +434,74 @@ void Jacobian::fullJacobian(const MultiBody& mb,
 	}
 }
 
+Eigen::MatrixXd Jacobian::expand(const MultiBody& mb,
+	const Eigen::Ref<const Eigen::MatrixXd>& jac) const
+{
+	Eigen::MatrixXd res = Eigen::MatrixXd::Zero(mb.nrDof(), mb.nrDof());
+	expandAdd(mb, jac, res);
+	return res;
+}
+
+void Jacobian::expandAdd(const MultiBody& mb,
+												 const Eigen::Ref<const Eigen::MatrixXd>& jac,
+												 Eigen::MatrixXd& res) const
+{
+	assert(res.cols() == mb.nrDof() && res.rows() == mb.nrDof());
+	int rowJac = 0;
+	int colJac = 0;
+	for(int i : jointsPath_)
+	{
+		colJac = 0;
+		for(int j : jointsPath_)
+		{
+			res.block(mb.jointPosInDof(i), mb.jointPosInDof(j), mb.joint(i).dof(), mb.joint(j).dof()).noalias()
+				+= jac.block(rowJac, colJac, mb.joint(i).dof(), mb.joint(j).dof());
+			colJac += mb.joint(j).dof();
+		}
+		rowJac += mb.joint(i).dof();
+	}
+}
+
+Blocks Jacobian::compactPath(const rbd::MultiBody& mb) const
+{
+	Blocks res;
+
+	int start_block = mb.jointPosInDof(jointsPath_[0]);
+	int len_block = mb.joint(jointsPath_[0]).dof();
+
+	int startJac = 0;
+
+	for (std::size_t j = 1;  j < jointsPath_.size(); ++j)
+	{
+		int i = jointsPath_[j];
+		int start = mb.jointPosInDof(i);
+
+		if(start != start_block + len_block)
+		{
+			res.emplace_back(start_block, startJac, len_block);
+			start_block = start;
+			startJac += len_block;
+			len_block = 0;
+		}
+		len_block += mb.joint(i).dof();
+	}
+	res.emplace_back(start_block, startJac, len_block);
+	return res;
+}
+
+void Jacobian::expandAdd(const Blocks& compactPath,
+												 const Eigen::Ref<const Eigen::MatrixXd>& jacMat,
+												 Eigen::MatrixXd& res) const
+{
+	for(const auto& b1 : compactPath)
+	{
+		for(const auto& b2 : compactPath)
+		{
+			res.block(b1.startDof, b2.startDof, b1.length, b2.length).noalias()
+				+= jacMat.block(b1.startJac, b2.startJac, b1.length, b2.length);
+		}
+	}
+}
 
 
 const Eigen::MatrixXd& Jacobian::sJacobian(
